streamlit_deploy_requirements.txt
streamlit
pandas
numpy
mlflow
scikit-learn
python-dotenv
catboost
xgboost
lightgbm